trial2
I'm having trouble making the other 3 cards change when it shuffles. you will have to load your own card pictures and just drag them into the same folder and add the file name . if you have a different approach for loading pictures and doing the random numbers please post it up ill try and work on this one. I copied all of Emil stuff also and put it on a grid.